backend/pyproject.toml
[tool.poetry]
name = "login-system-backend"
version = "0.1.0"
description = "認証付きTODOアプリ バックエンド"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
mangum = "^0.19.0"
uvicorn = { version = "^0.34.0", extras = ["standard"] }
boto3 = "^1.35.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
bcrypt = "^5.0.0"
pydantic-settings = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
moto = { version = "^5.0.0", extras = ["dynamodb"] }
httpx = "^0.28.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
boto3-stubs = { version = "^1.35.0", extras = ["dynamodb"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
